requirements.txt
python-dotenv>=1.0.0
duckdb>=0.10.0
google-generativeai>=0.7.0
openai>=1.0.0
anthropic>=0.40.0
pandas>=2.0.0
jieba>=0.42.1
pypdf>=4.0.0
python-docx>=1.1.0
chardet>=5.0.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.1.0
